% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{bernoulli_trial}
\alias{bernoulli_trial}
\title{Title}
\usage{
bernoulli_trial(outcome_set = c(0, 1), prob = 0.25)
}
\arguments{
\item{outcome_set}{set of two outcomes where second is considered a 'success'}

\item{prob}{probability of 'success', default is .25}
}
\value{
probability and outcome as 0 and 1
}
\description{
Title
}
\examples{
bernoulli_trial(.5)
bernoulli_trial(.6)
}
