% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{cross_trials}
\alias{cross_trials}
\title{Title}
\usage{
cross_trials(trial, num_trials = 2)
}
\arguments{
\item{num_trials}{}
}
\description{
Title
}
\examples{

bernoulli_trial() |>
  cross_trials(num_trials = 3)
}
