% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{fair_coin}
\alias{fair_coin}
\title{Title}
\usage{
fair_coin(outcome_set = c("tails", "heads"))
}
\arguments{
\item{prob}{probability of success, default is .25}
}
\value{
probability and outcome as 0 and 1
}
\description{
Title
}
\examples{
fair_coin()
fair_coin(outcome_set = 0:1)
fair_coin(outcome_set = c(FALSE, TRUE))
fair_coin(outcome_set = c("fish", "no fish"))
}
