% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{prize_wheel}
\alias{prize_wheel}
\title{Title}
\usage{
prize_wheel(back_story = TRUE)
}
\arguments{
\item{back_story}{}
}
\description{
Title
}
\examples{

prize_wheel()
prize_wheel(back_story = FALSE)

prize_wheel() |>
 trial_init() |>
 trial_advance()

prize_wheel(back_story = FALSE) |>
 trial_init() |>
 trial_advance()

}
