% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{prod_across}
\alias{prod_across}
\title{Title}
\usage{
prod_across(data, var_key = "prob")
}
\arguments{
\item{var_key}{}
}
\description{
Title
}
\examples{
library(dplyr)
library(magrittr)
bernoulli_trial(prob = .5) \%>\%
  add_trials() \%>\%
  add_trials() \%>\%
  add_trials(5) \%>\%
  .$out \%>\%
  sum_across() \%>\%
  prod_across() \%>\%
  group_by(global_outcome) \%>\%
  summarize(probs = sum(global_probs))


bernoulli_trial(prob = .5) \%>\%
  add_trials() \%>\%
  add_trials() \%>\%
  add_trials(5) \%>\%
  .$out \%>\%
  sum_across() \%>\%
  prod_across()
}
