% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{to_tsibble}
\alias{to_tsibble}
\title{Title}
\usage{
to_tsibble(trials)
}
\arguments{
\item{trials}{}
}
\description{
Title
}
\examples{
library(tidyverse)
bernoulli_trial() |>
  add_trials() |>
  to_tsibble()

bernoulli_trial(prob = .5) |>
  add_trials() |>
  add_trials() |>
  to_tsibble()  |>
  group_by(history)  |>
  summarize(hist_prob = prod(prob),
            count_successes = sum(outcome),
            paths = paste(outcome, collapse = ",")) |>
  arrange(count_successes) |>
  group_by(count_successes) |>
  summarize(count_prob = sum(hist_prob))

}
