% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{add_trials}
\alias{add_trials}
\title{Title}
\usage{
add_trials(trials, increment = 1)
}
\arguments{
\item{increment}{}
}
\description{
Title
}
\examples{
bernoulli_trial() |>
trial_init() |>
  add_trials() |>
  add_trials()

  bernoulli_trial() |>
  add_trials()

 fair_coin() |>
  add_trials(2)
}
