% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{trial_advance}
\alias{trial_advance}
\title{Title}
\usage{
trial_advance(trials, increment = 1, as_ts = FALSE)
}
\arguments{
\item{increment}{}
}
\description{
Title
}
\examples{
library(tidyverse)
bernoulli_trial() |>
trial_init() |>
  trial_advance()

bernoulli_trial() |>
trial_init() |>
  trial_advance(2)

bernoulli_trial() |>
trial_init() |>
  trial_advance(2, as_ts = TRUE)

bernoulli_trial() |>
trial_init() |>
  trial_advance() |>
  trial_advance()

}
