% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{trial_init}
\alias{trial_init}
\title{Title}
\usage{
trial_init(trial = NULL, prob = 0.25, as_ts = FALSE)
}
\arguments{
\item{trial}{}
}
\description{
Title
}
\examples{
bernoulli_trial() |>
trial_init()

bernoulli_trial() |>
trial_init(as_ts = TRUE)

trial_init() -> hi
}
