% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli-trial.R
\name{trial_init}
\alias{trial_init}
\title{Title}
\usage{
trial_init(trial = bernoulli_trial())
}
\arguments{
\item{trial}{}
}
\description{
Title
}
\examples{
bernoulli_trial() |>
trial_init()

trial_init()
}
